Json schema for blogs (#14)
* Refactor StickyCta component to update quizRoute for degree-completion path

* add schema for blog posts

* remove console log from admin bar

* ✨ (page.tsx): enhance metadata handling for blog post component
The changes improve the metadata generation for blog posts by providing fallback values. If the `meta.title` is not available, it now defaults to `postData.title`. Similarly, for the image, it falls back to `postFeaturedImage.url` or an empty string if no image is provided. This ensures that the blog post has meaningful metadata even when some data is missing, improving SEO and user experience.

---------

Co-authored-by: Smith, Bob <bobsmith@eab.com>