Qa (#20)
* fixed live preview for blog post page

* Render blog pages and category pages to serverside

* updated category component for fetching correct data from API

* updated category name for post url

* Json schema for blogs (#14)

* Refactor StickyCta component to update quizRoute for degree-completion path

* add schema for blog posts

* remove console log from admin bar

* ✨ (page.tsx): enhance metadata handling for blog post component
The changes improve the metadata generation for blog posts by providing fallback values. If the `meta.title` is not available, it now defaults to `postData.title`. Similarly, for the image, it falls back to `postFeaturedImage.url` or an empty string if no image is provided. This ensures that the blog post has meaningful metadata even when some data is missing, improving SEO and user experience.

---------

Co-authored-by: Smith, Bob <bobsmith@eab.com>

* 🪲 bugfix: destructure postFeaturedImage from postData

* update noindex to only affect children pages of business, healthcare, etc

* remove role under blog author name bc it shows "blogAuthor"

* 🔧 fix(serialize.tsx): add paragraph element with text-base class

* 🔧 fix(serialize.tsx): add paragraph element with text-base class

* 🔧 fix(serialize.tsx): add paragraph element with text-base class

* fixed live preview for blog post page

* 🔧 (package.json): resolve merge conflicts and clean up package.json
♻️ (page.tsx): remove unnecessary TypeScript ignore comments from BlogComponent

The merge conflicts in package.json have been resolved, ensuring that the file reflects the correct dependencies and scripts for the project. Additionally, unnecessary TypeScript ignore comments have been removed from the BlogComponent, improving code clarity and maintainability.

* Create npm-publish.yml

* Update npm-publish.yml

* initial add of sst

* merge main into qa & fix props for post page.tsx

* ✨ (package.json): add new script for SST development environment
♻️ (package.json): remove obsolete startsst script for clarity and consistency
A new script `dev:sst` is added to facilitate running the SST development environment. The obsolete `startsst` script is removed to reduce confusion and maintain a cleaner script section in the package.json file.

* remove workflow

* 📝 (mergeOpenGraph.ts): update siteName in defaultOpenGraph metadata for consistency
The siteName in the Open Graph metadata is updated from "Appily-Advance" to "Appily Advance" to ensure consistency in branding and improve readability. This change aligns the site name with the intended branding strategy, enhancing the overall presentation of the application on social media and other platforms.

* fixed live preview for blog post page

* 🔧 (package.json): resolve merge conflicts and clean up package.json
♻️ (page.tsx): remove unnecessary TypeScript ignore comments from BlogComponent

The merge conflicts in package.json have been resolved, ensuring that the file reflects the correct dependencies and scripts for the project. Additionally, unnecessary TypeScript ignore comments have been removed from the BlogComponent, improving code clarity and maintainability.

* Create npm-publish.yml

* Update npm-publish.yml

* initial add of sst

* merge main into qa & fix props for post page.tsx

* ✨ (package.json): add new script for SST development environment
♻️ (package.json): remove obsolete startsst script for clarity and consistency
A new script `dev:sst` is added to facilitate running the SST development environment. The obsolete `startsst` script is removed to reduce confusion and maintain a cleaner script section in the package.json file.

* remove workflow

* ✨ (middleware.ts): add pathname normalization for /blog paths to improve SEO and user experience
💡 (middleware.ts): add comments for better code clarity and understanding
The middleware now normalizes the pathname for any requests starting with "/blog" by converting it to lowercase and replacing spaces with hyphens. This change enhances SEO and provides a cleaner URL structure. Additionally, comments have been added to clarify the purpose of each code block, improving overall code readability.

* Remove unused code (#13)

* Remove unused files and code

* Refactor CarouselWithForm component and remove unused code

* 🔧 (GitHub workflows): update IAM role for GitHub actions to use dev role instead of production role

The IAM role used in multiple GitHub workflows has been changed from the production role to the development role. This change ensures that the workflows operate in a development environment, which is crucial for testing and development purposes without affecting the production environment.

---------

Co-authored-by: Komal Bambharoliya <KBambharoliya@eab.com>
Co-authored-by: Smith, Bob <bobsmith@eab.com>
Co-authored-by: Olea, Andrei <aolea@eab.com>