✨ (feature-branch.yaml): update workflow name to include 'Dev Branch' for clarity ✨ (feature-branch.yaml): add support for 'dev' branch in build preview workflow 🎣 (stage-for-production.yaml): create new workflow for automatic PR to production on merge to main ♻️ (stage.yaml): remove outdated staging deployment workflow
The workflow for building previews on feature branches is updated to include the 'dev' branch, enhancing clarity and functionality. A new workflow is introduced to automate the creation of pull requests to production when changes are merged into the main branch, streamlining the deployment process. The outdated staging deployment workflow is removed to reduce clutter and maintain focus on current practices.