first deploy (#23)
* update to gh

* remove pull request body

* update pull request workflow

* 📝 (stage-for-production.yaml): update pull request creation command to ensure proper formatting
The pull request title now includes a line continuation character to maintain clarity and prevent potential formatting issues. This change ensures that the title is correctly interpreted by the GitHub Actions workflow, improving the readability of the generated pull request.

* simplify

* ✨ (stage-for-production.yaml): add input for deployment reason in GitHub Actions workflow
♻️ (stage-for-production.yaml): refactor job name from create-pull-request-to-production to deploy for clarity
The workflow now includes an optional input for a deployment reason, allowing users to provide context for the deployment. This enhances transparency and documentation of changes being deployed. Additionally, renaming the job to "deploy" improves clarity regarding its purpose.

* 🔧 (dns.ts): update root domain from advance.appily.link to advance.appily.com for consistency
💡 (page.tsx): append 'test' to the description for testing purposes
The root domain has been updated to reflect the correct domain for the application. Additionally, the description in the degree completion page is modified to include 'test' to facilitate testing and ensure that changes can be easily identified during development.

* 📝 (github.ts, www.ts, sst-env.d.ts, tsconfig.json): comment out unused code and update configuration for clarity and maintainability
The changes involve commenting out unused code in `github.ts` to prevent execution while keeping it for future reference. In `www.ts`, the DNS and certificate settings are updated to ensure they are correctly configured for production. The `sst-env.d.ts` file is modified to define the `AppilyAdvanceSite` resource type, enhancing type safety. Lastly, the `tsconfig.json` file is updated to include `.next/types/**/*.ts` in the compilation process, ensuring that type definitions for Next.js are recognized. These changes improve code clarity, maintainability, and type safety across the project.