update workflow (#29)
* update to gh

* remove pull request body

* update pull request workflow

* 📝 (stage-for-production.yaml): update pull request creation command to ensure proper formatting
The pull request title now includes a line continuation character to maintain clarity and prevent potential formatting issues. This change ensures that the title is correctly interpreted by the GitHub Actions workflow, improving the readability of the generated pull request.

* simplify

* ✨ (stage-for-production.yaml): add input for deployment reason in GitHub Actions workflow
♻️ (stage-for-production.yaml): refactor job name from create-pull-request-to-production to deploy for clarity
The workflow now includes an optional input for a deployment reason, allowing users to provide context for the deployment. This enhances transparency and documentation of changes being deployed. Additionally, renaming the job to "deploy" improves clarity regarding its purpose.

* 🔧 (dns.ts): update root domain from advance.appily.link to advance.appily.com for consistency
💡 (page.tsx): append 'test' to the description for testing purposes
The root domain has been updated to reflect the correct domain for the application. Additionally, the description in the degree completion page is modified to include 'test' to facilitate testing and ensure that changes can be easily identified during development.

* 📝 (github.ts, www.ts, sst-env.d.ts, tsconfig.json): comment out unused code and update configuration for clarity and maintainability
The changes involve commenting out unused code in `github.ts` to prevent execution while keeping it for future reference. In `www.ts`, the DNS and certificate settings are updated to ensure they are correctly configured for production. The `sst-env.d.ts` file is modified to define the `AppilyAdvanceSite` resource type, enhancing type safety. Lastly, the `tsconfig.json` file is updated to include `.next/types/**/*.ts` in the compilation process, ensuring that type definitions for Next.js are recognized. These changes improve code clarity, maintainability, and type safety across the project.

* add permissions to production.yaml

* add change to dc

* remove pnpm build

* add change to dc (#24)

* add change to dc

* remove pnpm build

* add pnpm install

* add pnpm version

* add pnpm (#25)

* add environment

* hardcode env

* update arn role

* add logs to npx

* remove prod aws for testing

* remove prod aws for testing

* Update publish-to-aws.yaml

* 🔧 (publish-to-aws.yaml): update environment name to use dynamic input and re-enable production AWS credentials configuration
✨ (publish-to-aws.yaml): remove unnecessary print-logs option from deployment command
The environment name is now dynamically set using `${{ inputs.envName }}`, allowing for more flexible deployments. The production AWS credentials configuration step is re-enabled to ensure that the correct credentials are used when deploying to production. Additionally, the `--print-logs` option has been removed from the deployment command to streamline the output and focus on essential logs.

* update readme

* remove .png from infra folder

* Post author styling (#26)

* 🔨 refactor(page.client.tsx): clean up author bio section in PageClient component

🔨 refactor(PostPage.scss): update layout of author content wrapper to use grid instead of flex
🎨 style(PostPage.scss): adjust styling of h2 element within author content wrapper to be displayed in grid layout

* 🔨 refactor(page.tsx): clean up code formatting and remove unnecessary comments
🚀 feat(page.tsx): add support for category parameter in generateMetadata function
🐛 fix(page.tsx): fix category comparison logic in PostComponent

🔨 refactor(page.tsx): remove unnecessary code and fix conditional logic for category matching in PostComponent

* 🔇 chore(page.client.tsx): remove console.log statements from PageClient component

* 🔧 fix(styles): add styling for anchor tags in BlogPage.scss

* Posts inline blocks rendering bug (#27)

* Update production.yaml

---------

Co-authored-by: Andrettix <72987414+oleasky@users.noreply.github.com>